+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/MobileHeader/MobileHeader.scss
