+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/MobileHeader/index.ts
