+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/SublayoutHeader.scss
@@ -1,4 +1,4 @@
-@import 'shared';
+@import '../../../../styles/shared';
 
 /*
   Make the titlebar draggable for the desktop app
