+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/UserDropdown/UserDropdown.scss
