+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/UserDropdown/useCheckAuthenticatedAddresses.ts
