+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/index.ts
@@ -1,0 +1,3 @@
+import { SublayoutHeader } from './SublayoutHeader';
+
+export { SublayoutHeader };
